+++ db/updates/update_11.3-11.4.sql
@@ -1,1 +1,2 @@
 ALTER TABLE `settings` ADD `temp_sys` VARCHAR(255) NOT NULL DEFAULT '°C' AFTER `bs_theme`; 
+ALTER TABLE `ingredients` ADD `shelf_life` INT NOT NULL DEFAULT '0' AFTER `cid`;